Pointless cast

+++ src/main/java/svenhjol/charm/module/HuskImprovements.java
@@ -36,7 +36,7 @@ public class HuskImprovements extends MesonModule {
             ) {
                 World world = entity.getEntityWorld();
                 BlockPos pos = entity.getBlockPos();
-                int amount = ItemHelper.getAmountWithLooting(world.random, (int)maxDrops, lootingLevel, (float)lootingBoost);
+                int amount = ItemHelper.getAmountWithLooting(world.random, maxDrops, lootingLevel, (float)lootingBoost);
                 world.spawnEntity(new ItemEntity(world, pos.getX(), pos.getY(), pos.getZ(), new ItemStack(Blocks.SAND, amount)));
             }
             return ActionResult.PASS;
